+++ consumers/__init__.py
@@ -1,0 +1,1 @@
+"""Independent consumers of the native capability API."""